requirements.txt
python-dotenv~=0.17.1
python-telegram-bot~=13.5
torch~=1.9.0
numpy~=1.21.0
nltk~=3.6.2